LUN-208: rn kebabMenu, fix thread issues
